Renamed regex_t to onig_regex_t to prevent conflicts
It seems that with Swift and iOS 8, the typedef regex_t conflicts with
an existing typedef in the iOS8 SDK
(/Applications/Xcode6-Beta5.app/Contents/Developer/Platforms/iPhoneSimul
ator.platform/Developer/SDKs/iPhoneSimulator8.0.sdk/usr/include). With
Objective-C, it has no problems compiling, but there must be a bug in
the Swift compiler there is an issue which causes the names to conflict.